move files up a directory from buildtest/modules/ and update example files based on comments by @vsoch.
