+++ examples/moduletest.py
@@ -1,9 +1,4 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
-from buildtest.modules.util import Module
+from buildtest.module import Module
 
 mod_names = ["GCCcore", "Python"]
