+++ examples/saving_collection.py
@@ -1,8 +1,3 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
 from buildtest.modules.util import Module
 
 a = Module("zlib", debug=True)
